# 8293017: Improve hash calculation parallelism in jmod/jlink